% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_to_trig_mat}
\alias{array_to_trig_mat}
\title{Convert an array of covariance matrices to short lower triangular format}
\usage{
array_to_trig_mat(x, transposed = TRUE)
}
\description{
Convert an array of covariance matrices to short lower triangular format
}
\seealso{
\code{\link[=trig_mat_to_array]{trig_mat_to_array()}}, which is the reciprocal operation
}
