% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_average_overlap}
\alias{compute_average_overlap}
\title{Compute average overlap between components}
\usage{
compute_average_overlap(true_theta, k = length(true_theta$p))
}
\arguments{
\item{true_theta}{the parameters of the GMM}

\item{k}{the number of components}
}
\value{
the average overlap
}
\description{
Internally, it is the function MixSim::overlap which is used to generate an approximate pairwise
probability to wrongfully assign one component to another. Unfortunately, this function does not
the global overlap that we approximate there by averaging pairwise overlaps + compute the overlap
between two components accounting for their respective proportions in the mixture
}
\author{
Bastien CHASSAGNOL
}
