% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{compute_overlap_roots}
\alias{compute_overlap_roots}
\title{Compute the intersection points between two Gaussian mixtures (where one component goes above the other)}
\usage{
compute_overlap_roots(p1, p2, mu1, mu2, sigma1, sigma2)
}
\arguments{
\item{p1, p2, mu1, mu2, sigma1, sigma2}{parameters of the two Gaussian components}
}
\value{
the overlap roots
}
\description{
Compute the intersection points between two Gaussian mixtures (where one component goes above the other)
}
\author{
Bastien CHASSAGNOL
}
