% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{compute_parameters_complete_observations}
\alias{compute_parameters_complete_observations}
\title{Estimate the parameters in supervised case (all labels associated to the observations are available)}
\usage{
compute_parameters_complete_observations(simulated_distribution)
}
\arguments{
\item{simulated_distribution}{the object returned by generating a sample drawn from a GMM}
}
\value{
observed_estimated_theta a list of the estimated parameters in supervised case
}
\description{
Estimate the parameters in supervised case (all labels associated to the observations are available)
}
\seealso{
\code{\link{rnmix_skewed_with_outliers}}
}
\author{
Bastien CHASSAGNOL
}
