% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{compute_shannon_entropy}
\alias{compute_shannon_entropy}
\title{Compute the shannon entropy of a discrete distribution, normalised from 0 to 1 (equibalanced classes)}
\usage{
compute_shannon_entropy(ratios)
}
\arguments{
\item{ratios}{vector of the proportions of the mixture}
}
\value{
the entropy score
}
\description{
Compute the shannon entropy of a discrete distribution, normalised from 0 to 1 (equibalanced classes)
}
\author{
Bastien CHASSAGNOL
}
