% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cov_MLE}
\alias{cov_MLE}
\title{Compute the MLE covariance matrix estimate}
\usage{
cov_MLE(x)
}
\arguments{
\item{x}{the set of multi-observations, for which we would like to compute the covariance}
}
\description{
\code{\link[stats:cor]{stats::cov()}} function does not compute the MLE estimate of the covariance matrix,
but its unbiased version, with at the denominator, a \eqn{n-1} denominator that returns an unbiased estimate
of the covariance matrix.
}
\author{
Bastien CHASSAGNOL
}
