% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{estimate_supervised_univariate_GMM}
\alias{estimate_supervised_univariate_GMM}
\alias{estimate_supervised_multivariate_GMM}
\title{Estimate the parameters in supervised case (all labels associated to the observations are available)}
\usage{
estimate_supervised_univariate_GMM(x, s, s_outliers = s)

estimate_supervised_multivariate_GMM(x, s, k = 2)
}
\arguments{
\item{x}{the univariate or multivariate distribution of observed variables}

\item{s}{the index of the clusters, ranging from 1 to \eqn{k}, the number of clusters}

\item{s_outliers}{binary index, with 0 corresponding to outlying points}

\item{k}{the number of classes (by default, the number of unique values within the indicator vector)}
}
\value{
a list with three arguments, respectively the proportions, the means, and the variance / covariance
}
\description{
Estimate the parameters in supervised case (all labels associated to the observations are available)
}
\seealso{
\code{\link{simulate_univariate_GMM}}
}
\author{
Bastien CHASSAGNOL
}
