% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_theta_output}
\alias{format_theta_output}
\title{Format the estimated parameters}
\usage{
format_theta_output(theta)
}
\arguments{
\item{theta}{a list with 3 entries:
\itemize{
\item The proportions \code{p}: \eqn{p} of each component (must be included between 0 and 1, and sum to one overall)
\item The mean matrix \code{mu}: \eqn{\mathrm{\mu}=(\mu_{i,j}) \in \mathbb{R}^{n \times k}}, with each column
giving the mean values of the variables within a given component
\item The 3-dimensional covariance matrix array \code{Sigma}: \eqn{\mathrm{\Sigma}=(\Sigma_{i,j,l}) \in \mathbb{R}^{n \times n \times k}},
with each matrix \eqn{\Sigma_{..l}, l \in \{ 1, \ldots, k\}} storing the covariance matrix of a given component,
whose diagonal terms correspond to the variance of each variable, and off-terms diagonal elements return the covariance matrix
}}
}
\description{
Especially, we remove redundant pairwise correlations, and name adequatly each parameter
}
\author{
Bastien CHASSAGNOL
}
