% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{generate_ellipse}
\alias{generate_ellipse}
\title{Compute a set of points, in order to draw 95\% confidence intervals}
\usage{
generate_ellipse(theta, alpha = 0.05, npoints = 500)
}
\arguments{
\item{theta}{the list of the estimates returned by the EM algorithm}

\item{alpha}{the confidence interval}

\item{npoints}{the number of points used to generate the ellipes}
}
\value{
a tibble with \code{n} points associated to x and y locations associated to the \code{1-alpha} confidence interval
}
\description{
Compute a set of points, in order to draw 95\% confidence intervals
}
\author{
Bastien CHASSAGNOL
}
