% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_local_scores}
\alias{get_local_scores}
\title{Compute bias and mse for each parameter of a GMM distribution}
\usage{
get_local_scores(estimated_theta, true_theta)
}
\arguments{
\item{estimated_theta}{the list of the estimates returned by the EM algorithm}

\item{true_theta}{the true parameters}
}
\value{
a tibble with mean, standard deviation, bias and mse for each parameter
}
\description{
Compute bias and mse for each parameter of a GMM distribution
}
\author{
Bastien CHASSAGNOL
}
