% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_plot_limits}
\alias{get_plot_limits}
\title{Retrieve the limits of a package}
\usage{
get_plot_limits(plot)
}
\arguments{
\item{plot}{a ggplot object}
}
\value{
get limits within a plot
}
\description{
Retrieve the limits of a package
}
