% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hellinger}
\alias{hellinger}
\title{Compute the Hellinger distance}
\usage{
hellinger(mu1, Sigma1, mu2, Sigma2)
}
\value{
the Hellinger distance between two multivariate Gaussian distributions
}
\description{
Compute the Hellinger distance
}
