% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{initialize_em}
\alias{initialize_em}
\title{Global function to launch the initialisation step of the EM algorithm}
\usage{
initialize_em(
  x = NULL,
  k = 2,
  nstart = 10L,
  short_iter = 200,
  short_eps = 10^-2,
  prior_prob = 0.05,
  initialisation_algorithm = c("kmeans", "quantiles", "random", "hc", "small em",
    "rebmix"),
  ...
)
}
\arguments{
\item{x}{the vector of the observations}

\item{k}{the number of components}

\item{nstart}{the number of random restarts with kmeans, random and small EM method}

\item{short_iter, short_eps}{hyperparameters of the small EM method}

\item{prior_prob}{minimal uncertainity added to the minor components of each observation assigned by hierarchical clustering}

\item{initialisation_algorithm}{the choice of the initialisation method, between kmeans, quantiles, random, hc, small em and rebmix method}

\item{...}{additional hyperparameters supplied with some of the initialisation methods}
}
\value{
a list of the estimated parameters, ordered by increasing mean for identifiability issues
}
\description{
Global function to launch the initialisation step of the EM algorithm
}
\author{
Bastien CHASSAGNOL
}
