% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_integer}
\alias{is_integer}
\title{Test whether a number is an integer}
\usage{
is_integer(x)
}
\arguments{
\item{x}{(a numeric input of size 1 (real number))}
}
\value{
F if the Euclidean division by 1 does not return 0
}
\description{
Test whether a number is an integer
}
