% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_positive_definite}
\alias{is_positive_definite}
\title{Check the positive definiteness of a symmetric real matrix}
\usage{
is_positive_definite(sigma, tol = 1e-06)
}
\arguments{
\item{sigma}{a symmetric matrix with real values}

\item{tol}{the numerical maximal tolerance threshold,
to which an eigen value below it is considered negative}
}
\value{
a boolean, whether or not the matrix can be considered
positive definite or not
}
\description{
To do so, we compute the eigen values of the corresponding matrix with \code{eigen} function,
and if all of them are positive above a given threshold, then we return true.
}
\author{
Bastien CHASSAGNOL
}
