% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{logsumexp}
\alias{logsumexp}
\alias{predict_posterior_probability}
\title{Helper functions for the parameter estimation of GMMs}
\usage{
logsumexp(l)

predict_posterior_probability(x, estimated_theta)
}
\arguments{
\item{l}{a vector of numeric terms}

\item{x}{the vector of observed values, of size \eqn{n}}

\item{estimated_theta}{the estimated parameters}
}
\value{
a numeric scalar value, result of the previously described equation

a list with two elements:
\itemize{
\item the posterior probability matrix, \code{eta}: \eqn{\eta=(\eta_{i,j}) \in [0, 1]^{n \times k}}, with \eqn{\eta_{i,j}}
giving the posterior probability of observation \eqn{i} to belong to cluster \eqn{j}
\item \code{loglik} returns the expected log-likelihood of our experiment
}
}
\description{
These functions are small chunks of code designed to decompose the computation of the EM algorithm into simpler steps.
\code{logsumexp} returns the computation of equation \eqn{\log(\exp(sum(x)))}, avoiding numerical overflows
}
\section{Functions}{
\itemize{
\item \code{predict_posterior_probability}: \code{predict_posterior_probability} returns the expected probability for each observation
to belong to any of the \eqn{k} clusters set a priori, given the estimated parameters
}}

