% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{maha}
\alias{maha}
\title{The Maha distance}
\usage{
maha(mu, sigma)
}
\arguments{
\item{mu, sigma}{mean and covariance matrix of the distribution to which we have to compute the Mahalanobis distance}
}
\description{
The Maha distance
}
