% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_boxplots_parameters}
\alias{plot_boxplots_parameters}
\title{Plot the boxplot representation of the estimated parameters}
\usage{
plot_boxplots_parameters(
  distribution_parameters,
  true_theta,
  remove_outliers = T,
  size_tag = 4,
  num_col = length(true_theta$p)
)
}
\arguments{
\item{distribution_parameters}{the estimated bootstrap distributions}

\item{true_theta}{the true values of the parameters (with p, mu and sigma)}

\item{remove_outliers}{boolean: remove or not outlying estimates, by default set to True}

\item{size_tag}{in each facet, controls the size of the parameter emphasised}

\item{num_col}{How to organise the facets? by default according to the number of components}
}
\value{
boxplot_parameters a ggplot object representing the boxplot distributions of the estimates per package and initialisation package
}
\description{
Plot the boxplot representation of the estimated parameters
}
\author{
Bastien CHASSAGNOL
}
