% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_boxplots_parameters_univariate}
\alias{plot_boxplots_parameters_univariate}
\alias{plot_boxplots_parameters_bivariate}
\title{Plot the boxplot representation of the estimated parameters}
\usage{
plot_boxplots_parameters_univariate(
  distribution_parameters,
  p,
  mu,
  sigma,
  remove_outliers = T,
  size_tag = 8
)

plot_boxplots_parameters_bivariate(
  distribution_parameters,
  remove_outliers = T,
  size_tag = 8,
  num_col = NULL
)
}
\arguments{
\item{distribution_parameters}{the estimated bootstrap distributions}

\item{p, mu, sigma}{the true values of the parameters}

\item{with_outliers}{boolean: remove or not outlying estimates}
}
\value{
boxplot_parameters a ggplot object representing the boxplot distributions of the estimates per package and initialisation package
}
\description{
Plot the boxplot representation of the estimated parameters
}
\author{
Bastien CHASSAGNOL
}
