% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_correlation_Heatmap}
\alias{plot_correlation_Heatmap}
\title{Plot Correlation Heatmap}
\usage{
plot_correlation_Heatmap(distribution_parameters, row_km = 1)
}
\arguments{
\item{distribution_parameters}{A tibble with the distribution of the parameters,
in the long format}

\item{row_km}{The number of splits to carry on, by default set to 1}
}
\description{
For each configuration of parameter, compare the similarity between the packages
for each initialisation package, by computing the correlation between
their estimates
}
