% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_density_distribution}
\alias{plot_density_distribution}
\title{Plot the true density distributions}
\usage{
plot_density_distribution(
  proportions,
  mean_values,
  sigma_values,
  skewness_values
)
}
\arguments{
\item{proportions, mean_values, sigma_values, skewness_values}{parameters of the several components represented,
represented by a list of vector parameters}
}
\value{
density_plots a ggplot object representing the distributions of the density functions
}
\description{
Plot the true density distributions
}
\author{
Bastien CHASSAGNOL
}
