% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_ellipses_bivariate}
\alias{plot_ellipses_bivariate}
\title{Plot the confidence interval ellipses}
\usage{
plot_ellipses_bivariate(
  distribution_parameters,
  true_theta,
  alpha = 0.05,
  npoints = 500
)
}
\arguments{
\item{distribution_parameters}{the estimated bootstrap distributions}

\item{true_theta}{the true values of the parameters (with p, mu and sigma)}

\item{alpha}{the confidence interval}

\item{npoints}{the number of points used to generate the ellipses}
}
\value{
a ggplot object representing for each configuration of parameter, in each panel the confidence
intervals associated to the median of the estimate of each package
}
\description{
Plot the confidence interval ellipses
}
\author{
Bastien CHASSAGNOL
}
