% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_Hellinger}
\alias{plot_Hellinger}
\title{Plot boxplot of the Hellinger distances per component}
\usage{
plot_Hellinger(distribution_parameters, num_col = 2)
}
\arguments{
\item{distribution_parameters}{the estimated bootstrap distributions}
}
\value{
a ggplot object representing for each configuration of parameter, in each panel the
Hellinger distance for each package and initialisation method, respectively to each component
}
\description{
Plot boxplot of the Hellinger distances per component
}
\author{
Bastien CHASSAGNOL
}
