% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{plot_initialisation_time_computations}
\alias{plot_initialisation_time_computations}
\title{Plot the running time of the initialisation methods}
\usage{
plot_initialisation_time_computations(init_time_data)
}
\arguments{
\item{init_time_data}{the running time data}
}
\value{
initialisation_time_plots a ggplot object representing the running time curve distributions of the initialisation packages
}
\description{
Plot the running time of the initialisation methods
}
\author{
Bastien CHASSAGNOL
}
