% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_multi_parametrisation}
\alias{plot_multi_parametrisation}
\title{Generate the 14 configuration models}
\usage{
plot_multi_parametrisation(theta)
}
\arguments{
\item{theta}{the true values of the parameters (with p, mu and sigma)}
}
\value{
a ggplot object representing for each configuration of parameter, in each panel the confidence
intervals associated to the median of the estimate of each package
}
\description{
Generate the 14 configuration models
}
\author{
Bastien CHASSAGNOL
}
