% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_time_computations}
\alias{plot_time_computations}
\title{Plot the running time of the reviewed packages}
\usage{
plot_time_computations(time_data)
}
\arguments{
\item{time_data}{the running time data associated to the EM estimation performed by the reviewed packages}
}
\value{
time_plots a ggplot object representing the running time curve distributions of the reviewed packages
}
\description{
Plot the running time of the reviewed packages
}
\author{
Bastien CHASSAGNOL
}
