% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_univariate_normal_density_distribution}
\alias{plot_univariate_normal_density_distribution}
\alias{plot_bivariate_normal_density_distribution}
\title{Plot the true density distributions, in the univariate context}
\usage{
plot_univariate_normal_density_distribution(
  sigma_values,
  mean_values,
  proportions
)

plot_bivariate_normal_density_distribution(theta, n = 5000)
}
\arguments{
\item{proportions, mean_values, sigma_values}{parameters of the several components represented,
represented by a list of vector parameters}
}
\value{
density_plots a ggplot object representing the distributions of the density functions
}
\description{
Plot the true density distributions, in the univariate context
}
\author{
Bastien CHASSAGNOL
}
