% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{quadraticRoots}
\alias{quadraticRoots}
\title{Create quadratic formula function:}
\usage{
quadraticRoots(a, b, c)
}
\arguments{
\item{a}{quadratic parameter}

\item{b}{parameter of order 1}

\item{c}{intersect}
}
\value{
the roots of a quadratic function
}
\description{
Quadratic equation form of ax^2 + bx + c
}
