% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{rnmix_skewed_with_outliers}
\alias{rnmix_skewed_with_outliers}
\title{Generation of a mixture drawn from an univariate GMM, with possibility to add outliers and skewness}
\usage{
rnmix_skewed_with_outliers(
  n,
  theta = list(p = c(0.4, 0.6), mu = c(175, 165), sigma = c(10, 12), skew = c(0, 0)),
  prop_outliers = 0,
  interval = 2
)
}
\arguments{
\item{n}{the number of observations to be drawn}

\item{theta}{a list with 4 entries, corresponding to the true values of the parameters (proportion p, mean mu, deviation sigma, skewness skew)}

\item{prop_outliers, interval}{the proportion of outliers in the mixture, and its range respective to the 0.05 and 0.95 quantiles of the global distribution}
}
\value{
a list with the number of components k, the true parameters p, mu, sigma, skew, the observed variables x, the hidden observations s and an indicator of the outliers s_outliers
}
\description{
Generation of a mixture drawn from an univariate GMM, with possibility to add outliers and skewness
}
\author{
Bastien CHASSAGNOL
}
