% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{simulate_univariate_GMM}
\alias{simulate_univariate_GMM}
\alias{simulate_multivariate_GMM}
\title{Generation of a mixture drawn from an univariate GMM, with possibility to add outliers}
\usage{
simulate_univariate_GMM(
  theta = list(p = c(0.4, 0.6), mu = c(175, 165), sigma = c(10, 12)),
  n = 100,
  prop_outliers = 0,
  interval = 2
)

simulate_multivariate_GMM(theta, n = 500)
}
\arguments{
\item{theta}{a list with 3 entries:
\itemize{
\item The proportions \code{p}: \eqn{p} of each component (must be included between 0 and 1, and sum to one overall)
\item The mean matrix \code{mu}: \eqn{\mathrm{\mu}=(\mu_{i,j}) \in \mathbb{R}^{n \times k}}, with each column
giving the mean values of the variables within a given component
\item The 3-dimensional covariance matrix array \code{Sigma}: \eqn{\mathrm{\Sigma}=(\Sigma_{i,j,l}) \in \mathbb{R}^{n \times n \times k}}, with each matrix
\eqn{\Sigma_{..l}, l \in \{ 1, \ldots, k\}} storing the covariance matrix of a given component,
whose diagonal terms correspond to the variance of each variable, and off-terms diagonal elements return the covariance matrix
}}

\item{n}{the number of observations to be drawn}

\item{prop_outliers, interval}{first argument is the proportion of outliers, drawn from an uniform distribution whose
intervals are defined by the second argument twice the distance between the 0.05 and 0.95 quantile of the distribution}
}
\value{
Depending on the univariate or multivariate context:
\itemize{
\item a list with the number of components k, the true parameters p, mu, sigma,
the observed variables x, the hidden observations s and an indicator of the outliers s_outliers
\item a list with the number of components k, the true parameters p, mu, sigma,
the observations \code{X} living in space \eqn{\mathbb{R}^k}, the hidden indicator variables \code{s} (vector of size \eqn{n})
}
}
\description{
Generation of a mixture drawn from an univariate GMM, with possibility to add outliers
}
\details{
We do not implement the possibility of adding outliers in the multivariate context
}
\author{
Bastien CHASSAGNOL
}
