% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{trig_mat_to_array}
\alias{trig_mat_to_array}
\title{Convert condensed covariance matrix to large array format}
\usage{
trig_mat_to_array(x, transposed = TRUE, ...)
}
\arguments{
\item{x}{a covariance matrix, formatted as \eqn{\frac{p(p+1)}{2} \times k},
each column \eqn{j \in \{1, \ldots, k \}} describing the covariance matrix associated to component \eqn{j}}

\item{transposed}{if transposed, the input is of dimension \eqn{k \times \frac{p(p+1)}{2}},
each row \eqn{j} describing the covariance matrix associated to component \eqn{j}}

\item{...}{additional parameters to pass to function \code{\link[=vec2sym]{vec2sym()}}, enabling to deal with
distinct way to provide covariance information}
}
\value{
a large array, of dimension \eqn{p \times p \times k}: one array of covariance matrix,
on the third dimension, for each component
}
\description{
Shift from the lower triangular matrices \eqn{k \times \frac{p(p+1)}{2}} to large array, of dimension
\eqn{p \times p \times k}
}
