% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unformat_theta_output}
\alias{unformat_theta_output}
\title{Unformat the estimated parameters}
\usage{
unformat_theta_output(formatted_theta)
}
\arguments{
\item{formatted_theta}{a named vector, whose names refer to a given parameter and to an unique component}
}
\description{
Especially, we remove redundant pairwise correlations, and name adequatly each parameter
}
\seealso{
\code{\link[=format_theta_output]{format_theta_output()}}, the reciprocal operation
}
\author{
Bastien CHASSAGNOL
}
