% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vec2sym}
\alias{vec2sym}
\title{Convert vector to symmetric matrix}
\usage{
vec2sym(x, lower = TRUE, byrow = FALSE)
}
\arguments{
\item{x}{vector containing upper or lower triangular elements of a symmetric matrix.}

\item{lower}{x is from the lower triangle (default = \code{TRUE})}

\item{byrow}{the elements in x are ordered row-wise (default = \code{FALSE})}
}
\value{
Symmetric matrix
}
\description{
Reverse operation of \code{upper.tri} or \code{lower.tri}. Given a vector of
upper or lower triangular elements of a matrix
(optionally including the diagonal elements), returns the
corresponding symmetric matrix. The elements of the vector can be arranged by
column (the default) or by row.
}
\details{
Many thanks for the author of the \code{patr1ckm} package for providing us this method,
\href{https://rdrr.io/github/patr1ckm/patr1ckm/man/vec2sym.html}{Handy R functions},
enabling to revert an \code{upper.tri} or \code{lower.tri} operation.

Note that if x is a vector of the lower triagular elements given by column,
this is equivalent to the upper triangular elements given by row. Similarly, if
x is a vector of the lower triangular elements given by row, this is equivalent
to the upper triangular elements given by column.
}
\examples{
x <- c(1, 1, 1, 2, 2, 3)
check <- matrix(c(1, 1, 1, 1, 2, 2, 1, 2, 3), 3, 3)
identical(vec2sym(x, lower = TRUE, byrow = FALSE), check)
}
\author{
patr1ckm
}
