% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{array_to_trig_mat}
\alias{array_to_trig_mat}
\title{Convert an array of covariance matrices to short lower triangular format}
\usage{
array_to_trig_mat(x, transposed = TRUE)
}
\arguments{
\item{x}{an array, of dimensions \eqn{D \times D \times k},
each column \eqn{j \in \{1, \ldots, k \}} describing the covariance matrix associated to component \eqn{j}}

\item{transposed}{boolean, should we transpose the returned matrix}
}
\description{
Convert an array of covariance matrices to short lower triangular format
}
\seealso{
\code{\link[=trig_mat_to_array]{trig_mat_to_array()}}, which is the reciprocal operation
}
