% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixture.R
\name{emnmix}
\alias{emnmix}
\alias{em_Rmixmod}
\alias{em_EMCluster}
\alias{em_bgmm}
\alias{em_flexmix}
\alias{em_mixtools}
\alias{em_DCEM}
\alias{em_GMKMcharlie}
\alias{em_otrimle}
\alias{em_mixsmsn}
\title{Custom R implementation of the EM algorithm}
\usage{
emnmix(
  x,
  k,
  itmax = 5000,
  epsilon = 10^-12,
  nstart = 10L,
  start = NULL,
  initialisation_algorithm = "kmeans",
  ...
)

em_Rmixmod(
  x = x,
  k = 2,
  initialisation_algorithm = "kmeans",
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  ...
)

em_EMCluster(
  x = x,
  k = 2,
  initialisation_algorithm = "kmeans",
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  ...
)

em_bgmm(
  x = x,
  k = 2,
  itmax = 5000,
  epsilon = 10^-12,
  initialisation_algorithm = "kmeans",
  start = NULL,
  ...
)

em_flexmix(
  x = x,
  k = 2,
  itmax = 5000,
  epsilon = 10^-12,
  initialisation_algorithm = "kmeans",
  start = NULL,
  ...
)

em_mixtools(
  x = x,
  k = 2,
  initialisation_algorithm = "hc",
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  ...
)

em_DCEM(
  x = x,
  k = 2,
  initialisation_algorithm = "hc",
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  ...
)

em_GMKMcharlie(
  x = x,
  k = 2,
  initialisation_algorithm = "hc",
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  parallel = FALSE,
  ...
)

em_otrimle(
  x = x,
  k = 2,
  initialisation_algorithm = "hc",
  itmax = 5000,
  epsilon = 10^-12,
  ...
)

em_mixsmsn(
  x = x,
  k = 2,
  initialisation_algorithm = "hc",
  skew = rep(0, k),
  itmax = 5000,
  epsilon = 10^-12,
  start = NULL,
  ...
)
}
\arguments{
\item{x}{the vector of the observations}

\item{k}{the number of components}

\item{itmax}{the maximal number of iterations to reach the threshold}

\item{epsilon}{the criterion threshold considered as the tolerance between two consecutive log-likelihoods}

\item{start}{list of initial estimates provided by the user}

\item{initialisation_algorithm, nstart}{hyperparameters, when the user rather uses one of our implemented initialisation algorithms}

\item{...}{additional parameters for the reviewed packages}

\item{parallel}{only relevant for GMKMCharlie package which has a native parallell implementation (by default, takes half of the avalaible clusters)}

\item{skew}{the initial guess of the user on the skewness of the distribution (only relevent for em_mixsmn function)}
}
\value{
a list of the estimated parameters, ordered by increasing mean for identifiability issues
}
\description{
Custom R implementation of the EM algorithm
}
\section{Functions}{
\itemize{
\item \code{em_Rmixmod}: EM implementation with Rmixmod package

\item \code{em_EMCluster}: EM implementation with EMCluster package

\item \code{em_bgmm}: EM implementation with bgmm package

\item \code{em_flexmix}: EM implementation with flexmix package

\item \code{em_mixtools}: EM implementation with mixtools package

\item \code{em_DCEM}: EM implementation with DCEM package

\item \code{em_GMKMcharlie}: EM implementation with GMKMcharlie package

\item \code{em_otrimle}: EM implementation with otrimle package (designed to deal with outliers especially)

\item \code{em_mixsmsn}: EM implementation with mixsmsn package (designed to deal with skewed GMMs especially)
}}

\author{
Bastien CHASSAGNOL
}
