% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_boxplots_parameters}
\alias{plot_boxplots_parameters}
\title{Plot the boxplot representation of the estimated parameters}
\usage{
plot_boxplots_parameters(
  distribution_parameters,
  p,
  mu,
  sigma,
  with_outliers = FALSE
)
}
\arguments{
\item{distribution_parameters}{the estimated bootstrap distributions}

\item{p, mu, sigma}{the true values of the parameters}

\item{with_outliers}{boolean: remove or not outlying estimates}
}
\value{
boxplot_parameters a ggplot object representing the boxplot distributions of the estimates per package and initialisation algorithm
}
\description{
Plot the boxplot representation of the estimated parameters
}
\author{
Bastien CHASSAGNOL
}
