% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_correlation_Heatmap}
\alias{plot_correlation_Heatmap}
\title{Plot Correlation Heatmap}
\usage{
plot_correlation_Heatmap(distribution_parameters)
}
\arguments{
\item{distribution_parameters}{A tibble with the distribution of the parameters,
in the long format}
}
\description{
For each configuration of parameter, compare the similarity between the algorithms
for each initialisation algorithm, by computing the correlation between
their estimates
}
