% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualisation.R
\name{plot_univariate_normal_density_distribution}
\alias{plot_univariate_normal_density_distribution}
\alias{plot_bivariate_normal_density_distribution}
\title{Plot the true density distributions, in the univariate context}
\usage{
plot_univariate_normal_density_distribution(
  true_theta,
  nobservations = 1000,
  k = length(true_theta$p)
)

plot_bivariate_normal_density_distribution(
  true_theta,
  nobservations = 1000,
  k = length(true_theta$p)
)
}
\arguments{
\item{true_theta}{the true values of the parameters (with p, mu and sigma)}

\item{nobservations}{the number of bins used to regenerate the global mixture distribution}

\item{k}{the number of components, by default, the number of ratio elements stored in item p of true_theta}
}
\value{
density_plots a ggplot object representing the distributions of empirical univariate density functions
}
\description{
Plot the true density distributions, in the univariate context
}
\author{
Bastien CHASSAGNOL
}
